requirements.txt
numpy
trimesh
pillow
torch
pygfx
wgpu
pygltflib
scipy
